Fix bug with tableDescription call and failFast
If any table description was in bad shape, it will throw an exception right
away. It wouldn't matter if fail fast was true or not, it won't even go
into the controller. The idea was always that it won't matter if there
is any failings or not, the process should end ok by default and only
when the user don't want that to happen and prefer to fail, then should
fail.