Fix error while reading 'Á' & 'Í' chars from file
When reading a plain text file from system that contains either an Á or
an Í the read_file method fails withouth warnings or exceptions. Every
other capital letter or accent is read properly, but those aren't. It's
even more strange that, when running from a test env like Intellij test
pass, but doing them from Maven fails. Forcing the enconding to UTF-8
seems to fix this issue.