Fix bug with different spacings in Decimal type conversion
Using a split method on string line is good if you can maintain
consistency in the input, but there are some time that consistency goes
out of the window. For example, when parsing a Decimal with a space
between the comma and the second value, the method will split in the
comma and thus everything goes wrong.